default dtype is f16 except the rmsnorm

+++ src/main.rs
@@ -1,4 +1,5 @@
 use anyhow::Result;
+use dfdx::dtypes::f16;
 use dfdx::prelude::*;
 use rand::prelude::{SeedableRng, StdRng};
 
@@ -80,7 +81,7 @@ fn main() -> Result<()> {
 
     let cfg = ConfigV2 {};
 
-    let m = Mistral::<f32, _, _>::load_model(cfg, &dev, &loader)?;
+    let m = Mistral::<f16, _, _>::load_model(cfg, &dev, &loader)?;
 
     let gen_opt = GenerateOption {
         use_cache: !args.disable_cache,
